# fix: decimal conversion looses value on lower precision