--- a/arrow-cast/src/cast/decimal.rs
+++ b/arrow-cast/src/cast/decimal.rs
@@ -112,8 +112,19 @@ where
     };
 
     Ok(match cast_options.safe {
-        true => array.unary_opt(f),
-        false => array.try_unary(|x| f(x).ok_or_else(|| error(x)))?,
+        true => {
+            array.unary_opt(|x| {
+                 f(x).filter(|v| O::is_valid_decimal_precision(*v, output_precision))
+            })
+        }
+        false => {
+            array.try_unary(|x| {
+                f(x).ok_or_else(|| error(x))
+                    .and_then(|v|{
+                        O:: validate_decimal_precision(v, output_precision).map(|_| v)
+                    })
+            })?
+        }
     })
 }
 
@@ -135,11 +146,23 @@ where
         .unwrap()
         .pow_checked((output_scale - input_scale) as u32)?;
 
-    let f = |x| O::Native::from_decimal(x).and_then(|x| x.mul_checked(mul).ok());
+    let f = |x|
+        O::Native::from_decimal(x).and_then(|x| x.mul_checked(mul).ok());
 
     Ok(match cast_options.safe {
-        true => array.unary_opt(f),
-        false => array.try_unary(|x| f(x).ok_or_else(|| error(x)))?,
+        true => {
+            array.unary_opt(|x| {
+                f(x).filter(|v| O::is_valid_decimal_precision(*v, output_precision))
+            })
+        }
+        false => {
+            array.try_unary(|x| {
+                 f(x).ok_or_else(|| error(x))
+                    .and_then(|v|{
+                        O:: validate_decimal_precision(v, output_precision).map(|_| v)
+                })
+            })?
+        }
     })
 }
 
@@ -156,19 +179,9 @@ where
     T::Native: DecimalCast + ArrowNativeTypeOp,
 {
     let array: PrimitiveArray<T> = match input_scale.cmp(&output_scale) {
-        Ordering::Equal => {
-            // the scale doesn't change, the native value don't need to be changed
-            array.clone()
-        }
-        Ordering::Greater => convert_to_smaller_scale_decimal::<T, T>(
-            array,
-            input_scale,
-            output_precision,
-            output_scale,
-            cast_options,
-        )?,
-        Ordering::Less => {
-            // input_scale < output_scale
+        Ordering::Equal | Ordering::Less => {
+            // input_scale <= output_scale
+            // the scale doesn't change, but precision may change and cause overflow
             convert_to_bigger_or_equal_scale_decimal::<T, T>(
                 array,
                 input_scale,
@@ -177,6 +190,13 @@ where
                 cast_options,
             )?
         }
+        Ordering::Greater => convert_to_smaller_scale_decimal::<T, T>(
+            array,
+            input_scale,
+            output_precision,
+            output_scale,
+            cast_options,
+        )?
     };
 
     Ok(Arc::new(array.with_precision_and_scale(