@andygrove 2024-12-04
There is a regression in `test_cast_decimal128_to_decimal256_negative`. The new validation check is correctly throwing an error, but it looks like we also need to add this validation when creating decimal arrays since the current test is creating invalid arrays before the cast:

```rust
let array = vec![Some(i128::MAX), Some(i128::MIN)];
let input_decimal_array = create_decimal_array(array, 10, 3).unwrap();
```

I would expect this to fail, so we probably need to add the same validation there.