@andygrove 2024-12-04
> Have we run any benchmarks (I'm not sure if any actually exist) to confirm this doesn't significantly regress performance.
> 
> It seems unfortunate to be always performing overflow checks, when in many cases it should be possible to prove that precision overflow can't occur and need not be checked for

I'll create a separate PR (probably tomorrow) to add some criterion benchmarks